fix: move layering concept into position
