+++ ryot/src/bevy_ryot/appearances.rs
@@ -3,7 +3,7 @@
 //! This file contains the information needed to load sprites and other content.
 use crate::appearances::{self, FixedFrameGroup};
 use crate::appearances::{AppearanceFlags, Appearances, FrameGroup};
-use crate::prelude::{drawing::*, *};
+use crate::prelude::*;
 use bevy::asset::io::Reader;
 use bevy::asset::{Asset, AssetLoader, AsyncReadExt, BoxedFuture, LoadContext};
 use bevy::prelude::*;
@@ -12,6 +12,8 @@ use bevy::utils::HashMap;
 use prost::Message;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
+
+use self::position::Layer;
 
 /// A plugin to register the Appearance asset and its loader.
 pub struct AppearanceAssetPlugin;
