+++ ryot/src/bevy_ryot/drawing/mod.rs
@@ -1,6 +1,6 @@
 use crate::appearances::{self, FixedFrameGroup};
 use crate::bevy_ryot::{AppearanceDescriptor, InternalContentState};
-use crate::position::TilePosition;
+use crate::position::{Layer, TilePosition};
 use bevy::prelude::*;
 use bevy::utils::HashMap;
 
@@ -171,55 +171,6 @@ impl DetailLevel {
             Self::GroundBottom => vec![Layer::Bottom, Layer::Ground],
             Self::GroundOnly => vec![Layer::Ground],
             Self::None => vec![],
-        }
-    }
-}
-
-/// This enum defines the layers that composes a game.
-/// The base layers are defined in the enum and custom layers can be added.
-/// The layers are used to define the rendering order of the tiles.
-/// There are two types of layers: dynamic and static.
-///     - Dynamic layers are used for elements that can change their rendering order
-///     depending on their attributes like position or floor. E.g.: creatures, items, effects.
-///    - Static layers are used for elements that have a fixed rendering order and are always
-///    rendered on top of dynamic layers. E.g.: mouse, grid, ui elements.
-///
-/// Static layers are separated from dynamic layers by the StaticLowerBound layer.
-#[derive(Eq, Hash, PartialEq, Debug, Copy, Clone, Default, Reflect, Component)]
-pub enum Layer {
-    #[default]
-    Items,
-    Top,
-    Bottom,
-    Ground,
-    Creature,
-    Effect,
-    StaticLowerBound,
-    Cursor,
-    Custom(i32),
-}
-
-impl Layer {
-    /// Z is calculated based on floor position and layer.
-    /// The base Z is floor * 100 and the layer adds an offset.
-    /// The offset is used to calculate the rendering order of the tile.
-    /// Tiles with higher Z are rendered on top of tiles with lower Z.
-    /// The tile Z for the game is always the floor (Z / 100).floor().
-    ///
-    /// We leave a gap of 10 between layers to allow for more layers to be added
-    /// in the future and to make it possible to place custom layers between
-    /// the default ones.
-    pub fn z(&self) -> i32 {
-        match self {
-            Self::Top => 90,
-            Self::Bottom => 80,
-            Self::Effect => 60,
-            Self::Creature => 40,
-            Self::Items => 20,
-            Self::Ground => 0,
-            Self::StaticLowerBound => 900,
-            Self::Cursor => 999,
-            Self::Custom(z) => *z,
         }
     }
 }
